More rename from Fluid to Fluid. Removed common project
