+++ Source/Bars.Mvvm.FluentApi.sln
@@ -22,8 +22,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		..\Version.json = ..\Version.json
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentApi.Common", "FluentApi.Common\FluentApi.Common.csproj", "{09776564-1CA1-4AC2-8036-1FC36FE9CBC7}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{32D7E9E8-71A2-4896-A862-19BEFC0F54B6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{115177F1-0D7E-4995-91AA-C8EFCC491FD8}"
@@ -63,10 +61,6 @@ Global
 		{66660992-6D45-4A59-AE3C-B943961327DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{66660992-6D45-4A59-AE3C-B943961327DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{66660992-6D45-4A59-AE3C-B943961327DD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{09776564-1CA1-4AC2-8036-1FC36FE9CBC7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{09776564-1CA1-4AC2-8036-1FC36FE9CBC7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{09776564-1CA1-4AC2-8036-1FC36FE9CBC7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{09776564-1CA1-4AC2-8036-1FC36FE9CBC7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{94DDD2DE-D0D8-49B5-97ED-8405F790429B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{94DDD2DE-D0D8-49B5-97ED-8405F790429B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94DDD2DE-D0D8-49B5-97ED-8405F790429B}.Release|Any CPU.ActiveCfg = Release|Any CPU
