--- Source/FluentApi.Common/FluentApi.Common.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <RootNamespace>ClassLibrary1</RootNamespace>
-        <LangVersion>latest</LangVersion>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-    <PropertyGroup>
-        <PackageScribanIncludeSource>false</PackageScribanIncludeSource>
-        <DefineConstants>SCRIBAN_NO_ASYNC;SCRIBAN_PUBLIC</DefineConstants>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="Microsoft.CodeAnalysis.Common" />
-        <PackageReference Include="Scriban" IncludeAssets="Build" />
-    </ItemGroup>
-    <PropertyGroup>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-</Project>
