+++ Source/FluentApi.Generator.Wpf.Tests/ConvenienceApi/RibbonViewModelShould.cs
@@ -3,7 +3,7 @@ using ActiproSoftware.Windows.Controls.Bars.Mvvm;
 using FluentAssertions;
 using Xunit;
 
-namespace Bars.Mvvm.FluentApi.Generator.Wpf.Test;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions.Wpf.Test;
 /// <summary>
 /// Verifies the output of the generator, and assert against source classes.
 /// </summary>
