+++ Source/FluentApi.Generator.Wpf.Tests/FluentApi.Generator.Wpf.Tests.csproj
@@ -11,7 +11,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\FluentApi.Common\FluentApi.Common.csproj" />
         <ProjectReference Include="..\FluentApi.Generator\FluentApi.Generator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
     </ItemGroup>
