+++ Source/FluentApi.Generator/Extensions/ActiproTypeExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 
-namespace Bars.Mvvm.FluentApi.Common;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions;
 
 /// <summary>
 /// Extension methods for <see cref="INamedTypeSymbol"/>.
