+++ Source/FluentApi.Generator/FluidCodeGenerator/FluidExtensionCodeGenerator.cs
@@ -1,15 +1,15 @@
-﻿using Bars.Mvvm.FluentApi.Common;
+﻿using Bars.Mvvm.FluentApi.Generator.Extensions;
+using Bars.Mvvm.FluentApi.Generator.Extensions.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Threading;
-using Bars.Mvvm.FluentApi.Generator.Models;
 using Microsoft.CodeAnalysis;
 using Scriban;
 
-namespace Bars.Mvvm.FluentApi.Generator;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions;
 
 /// <summary>
 /// This generator is responsible for generating Fluid API for Bars MVVM ViewModels for easy configuration.
