+++ Source/FluentApi.Generator/FluidCodeGenerator/TemplateModels/IExtensionTemplateModel.cs
@@ -1,4 +1,4 @@
-namespace Bars.Mvvm.FluentApi.Generator.Models;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions.Models;
 
 /// <summary>
 /// Marker template model interface for all property template models.
