+++ Source/FluentApi.Generator/FluidCodeGenerator/TemplateModels/ObservableCollectionTemplateModel.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 
-namespace Bars.Mvvm.FluentApi.Generator.Models;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions.Models;
 
 /// <summary>
 /// Model used for generating ObservableCollection properties.
