+++ Source/FluentApi.Generator/Models/EquatableList.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Bars.Mvvm.FluentApi.Common;
+namespace Bars.Mvvm.FluentApi.Generator.Models;
 
 /// <summary>
 /// EquitableList is a custom list that implements IEquatable to allow for value-based equality checks..
