+++ Source/FluentApi.Wpf.Sample/AdditionalExtensions.cs
@@ -2,7 +2,7 @@ using ActiproSoftware.Windows.Controls.Bars.Mvvm;
 using MahApps.Metro.IconPacks;
 using System.Windows.Media;
 
-namespace Bars.Mvvm.FluentApi.Generator.Sample;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions.Sample;
 
 public static class AdditionalExtensions
 {
