+++ Source/FluentApi.Wpf.Sample/MainWindow.xaml
@@ -1,4 +1,4 @@
-﻿<bars:RibbonWindow x:Class="Bars.Mvvm.FluentApi.Generator.Sample.MainWindow"
+﻿<bars:RibbonWindow x:Class="Bars.Mvvm.FluentApi.Generator.Extensions.Sample.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
