+++ Source/FluentApi.Wpf.Sample/SampleMenu/SampleBarManager.cs
@@ -1,7 +1,7 @@
 using ActiproSoftware.Windows.Controls.Bars;
 using ActiproSoftware.Windows.Controls.Bars.Mvvm;
 
-namespace Bars.Mvvm.FluentApi.Generator.Sample;
+namespace Bars.Mvvm.FluentApi.Generator.Extensions.Sample;
 
 public class SampleBarManager
 {
